[studio] select synchronization between table/grahpic

+++ ProlineStudio-ResultExplorer/src/main/java/fr/proline/studio/rsmexplorer/gui/ptm/ViewSetting.java
@@ -18,7 +18,7 @@ public class ViewSetting {
     /**
      * the width to draw an Amino acid
      */
-    public static double WIDTH_AA;
+    public static int WIDTH_AA;
     public static int HEIGHT_AA;
 
     public static int BORDER_GAP = 5;
